style: fix linter issues and data location
