+++ src/authz-module/data/api.ts
@@ -1,7 +1,7 @@
 import { getAuthenticatedHttpClient } from '@edx/frontend-platform/auth';
-import { LibraryMetadata, TeamMember } from '@src/authz-module/constants';
+import { LibraryMetadata, TeamMember } from '@src/types';
 import { camelCaseObject } from '@edx/frontend-platform';
-import { getApiUrl, getStudioApiUrl } from '@src/helpers/utils';
+import { getApiUrl, getStudioApiUrl } from '@src/data/utils';
 
 export interface GetTeamMembersResponse {
   members: TeamMember[];
@@ -9,8 +9,8 @@ export interface GetTeamMembersResponse {
 }
 
 // TODO: replece api path once is created
-export const getTeamMembers = async (libraryId: string): Promise<TeamMember[]> => {
-  const { data } = await getAuthenticatedHttpClient().get(getApiUrl(`/api/authz/v1/roles/users?scope=${libraryId}`));
+export const getTeamMembers = async (object: string): Promise<TeamMember[]> => {
+  const { data } = await getAuthenticatedHttpClient().get(getApiUrl(`/api/authz/v1/roles/users?scope=${object}`));
   return camelCaseObject(data.results);
 };
