+++ src/components/LoadingPage.tsx
@@ -1,14 +1,14 @@
 import { Spinner, Container } from '@openedx/paragon';
 
-const LoadingPage= () => {
+const LoadingPage = () => {
   return (
     <Container className="d-flex vh-100" data-testid="loading-page">
-          <Spinner
-            variant="primary"
-            animation="border"
-            role="status"
-            className="mb-3 m-auto"
-          />
+      <Spinner
+        variant="primary"
+        animation="border"
+        role="status"
+        className="mb-3 m-auto"
+      />
     </Container>
   );
 };
