+++ src/types.ts
@@ -1,0 +1,31 @@
+export interface PermissionValidationRequest {
+    action: string;
+    object?: string;
+    scope?: string;
+};
+
+export interface PermissionValidationResponse extends PermissionValidationRequest{
+    allowed: boolean;
+};
+
+// Libraries AuthZ types
+export interface TeamMember {
+  displayName: string;
+  username: string;
+  email: string;
+  roles: string[];
+}
+
+export interface LibraryMetadata {
+  id: string;
+  org: string;
+  title: string;
+  slug: string;
+}
+
+// Paragon table type
+export interface TableCellValue<T> {
+  row: {
+    original: T;
+  };
+}
